+++ src/unix/lwt_unix_unix.h
@@ -41,127 +41,6 @@
 #include <sys/wait.h>
 #include <unistd.h>
 
-/* +-----------------------------------------------------------------+
-   | Test for readability/writability                                |
-   +-----------------------------------------------------------------+ */
-
-CAMLprim value lwt_unix_readable(value fd)
-{
-    struct pollfd pollfd;
-    pollfd.fd = Int_val(fd);
-    pollfd.events = POLLIN;
-    pollfd.revents = 0;
-    if (poll(&pollfd, 1, 0) < 0) uerror("readable", Nothing);
-    return (Val_bool(pollfd.revents & POLLIN));
-}
-
-CAMLprim value lwt_unix_writable(value fd)
-{
-    struct pollfd pollfd;
-    pollfd.fd = Int_val(fd);
-    pollfd.events = POLLOUT;
-    pollfd.revents = 0;
-    if (poll(&pollfd, 1, 0) < 0) uerror("writable", Nothing);
-    return (Val_bool(pollfd.revents & POLLOUT));
-}
-
-/* +-----------------------------------------------------------------+
-   | Memory mapped files                                             |
-   +-----------------------------------------------------------------+ */
-
-static int advise_table[] = {
-    MADV_NORMAL, MADV_RANDOM, MADV_SEQUENTIAL, MADV_WILLNEED, MADV_DONTNEED,
-};
-
-CAMLprim value lwt_unix_madvise(value val_buffer, value val_offset,
-                                value val_length, value val_advice)
-{
-    int ret =
-        madvise((char *)Caml_ba_data_val(val_buffer) + Long_val(val_offset),
-                Long_val(val_length), advise_table[Int_val(val_advice)]);
-    if (ret == -1) uerror("madvise", Nothing);
-    return Val_unit;
-}
-
-CAMLprim value lwt_unix_get_page_size(value Unit)
-{
-    long page_size = sysconf(_SC_PAGESIZE);
-    if (page_size < 0) page_size = 4096;
-    return Val_long(page_size);
-}
-
-#ifdef __CYGWIN__
-LWT_NOT_AVAILABLE4(unix_mincore)
-#else
-
-#ifdef HAVE_BSD_MINCORE
-#define MINCORE_VECTOR_TYPE char
-#else
-#define MINCORE_VECTOR_TYPE unsigned char
-#endif
-
-CAMLprim value lwt_unix_mincore(value val_buffer, value val_offset,
-                                value val_length, value val_states)
-{
-    long len = Wosize_val(val_states);
-    MINCORE_VECTOR_TYPE vec[len];
-    mincore((char *)Caml_ba_data_val(val_buffer) + Long_val(val_offset),
-            Long_val(val_length), vec);
-    long i;
-    for (i = 0; i < len; i++) Field(val_states, i) = Val_bool(vec[i] & 1);
-    return Val_unit;
-}
-
-#undef MINCORE_VECTOR_TYPE
-
-#endif
-
-/* +-----------------------------------------------------------------+
-   | read/write                                                      |
-   +-----------------------------------------------------------------+ */
-
-CAMLprim value lwt_unix_read(value val_fd, value val_buf, value val_ofs,
-                             value val_len)
-{
-    long ret;
-    ret = read(Int_val(val_fd), &Byte(String_val(val_buf), Long_val(val_ofs)),
-               Long_val(val_len));
-    if (ret == -1) uerror("read", Nothing);
-    return Val_long(ret);
-}
-
-CAMLprim value lwt_unix_bytes_read(value val_fd, value val_buf, value val_ofs,
-                                   value val_len)
-{
-    long ret;
-    ret = read(Int_val(val_fd),
-               (char *)Caml_ba_array_val(val_buf)->data + Long_val(val_ofs),
-               Long_val(val_len));
-    if (ret == -1) uerror("read", Nothing);
-    return Val_long(ret);
-}
-
-CAMLprim value lwt_unix_write(value val_fd, value val_buf, value val_ofs,
-                              value val_len)
-{
-    long ret;
-    ret = write(Int_val(val_fd), &Byte(String_val(val_buf), Long_val(val_ofs)),
-                Long_val(val_len));
-    if (ret == -1) uerror("write", Nothing);
-    return Val_long(ret);
-}
-
-CAMLprim value lwt_unix_bytes_write(value val_fd, value val_buf, value val_ofs,
-                                    value val_len)
-{
-    long ret;
-    ret = write(Int_val(val_fd),
-                (char *)Caml_ba_array_val(val_buf)->data + Long_val(val_ofs),
-                Long_val(val_len));
-    if (ret == -1) uerror("write", Nothing);
-    return Val_long(ret);
-}
-
 /* readv, writev */
 
 /* For blocking readv calls, arrays of this struct associate temporary buffers
