+++ src/unix/lwt_unix_windows.h
@@ -26,153 +26,6 @@ CAMLprim value lwt_unix_is_socket(value fd)
     return (Val_bool(Descr_kind_val(fd) == KIND_SOCKET));
 }
 
-CAMLprim value lwt_unix_write(value fd, value buf, value vofs, value vlen)
-{
-    intnat ofs, len, written;
-    DWORD numbytes, numwritten;
-    DWORD err = 0;
-
-    Begin_root(buf);
-    ofs = Long_val(vofs);
-    len = Long_val(vlen);
-    written = 0;
-    if (len > 0) {
-        numbytes = len;
-        if (Descr_kind_val(fd) == KIND_SOCKET) {
-            int ret;
-            SOCKET s = Socket_val(fd);
-            ret = send(s, &Byte(buf, ofs), numbytes, 0);
-            if (ret == SOCKET_ERROR) err = WSAGetLastError();
-            numwritten = ret;
-        } else {
-            HANDLE h = Handle_val(fd);
-            if (!WriteFile(h, &Byte(buf, ofs), numbytes, &numwritten, NULL))
-                err = GetLastError();
-        }
-        if (err) {
-            win32_maperr(err);
-            uerror("write", Nothing);
-        }
-        written = numwritten;
-    }
-    End_roots();
-    return Val_long(written);
-}
-
-CAMLprim value lwt_unix_bytes_write(value fd, value buf, value vofs, value vlen)
-{
-    intnat ofs, len, written;
-    DWORD numbytes, numwritten;
-    DWORD err = 0;
-
-    Begin_root(buf);
-    ofs = Long_val(vofs);
-    len = Long_val(vlen);
-    written = 0;
-    if (len > 0) {
-        numbytes = len;
-        if (Descr_kind_val(fd) == KIND_SOCKET) {
-            int ret;
-            SOCKET s = Socket_val(fd);
-            ret = send(s, (char *)Caml_ba_array_val(buf)->data + ofs, numbytes,
-                       0);
-            if (ret == SOCKET_ERROR) err = WSAGetLastError();
-            numwritten = ret;
-        } else {
-            HANDLE h = Handle_val(fd);
-            if (!WriteFile(h, (char *)Caml_ba_array_val(buf)->data + ofs,
-                           numbytes, &numwritten, NULL))
-                err = GetLastError();
-        }
-        if (err) {
-            win32_maperr(err);
-            uerror("write", Nothing);
-        }
-        written = numwritten;
-    }
-    End_roots();
-    return Val_long(written);
-}
-
-CAMLprim value lwt_unix_read(value fd, value buf, value vofs, value vlen)
-{
-    intnat ofs, len, written;
-    DWORD numbytes, numwritten;
-    DWORD err = 0;
-
-    Begin_root(buf);
-    ofs = Long_val(vofs);
-    len = Long_val(vlen);
-    written = 0;
-    if (len > 0) {
-        numbytes = len;
-        if (Descr_kind_val(fd) == KIND_SOCKET) {
-            int ret;
-            SOCKET s = Socket_val(fd);
-            ret = recv(s, &Byte(buf, ofs), numbytes, 0);
-            if (ret == SOCKET_ERROR) err = WSAGetLastError();
-            numwritten = ret;
-        } else {
-            HANDLE h = Handle_val(fd);
-            if (!ReadFile(h, &Byte(buf, ofs), numbytes, &numwritten, NULL))
-                err = GetLastError();
-        }
-        if (err) {
-            win32_maperr(err);
-            uerror("write", Nothing);
-        }
-        written = numwritten;
-    }
-    End_roots();
-    return Val_long(written);
-}
-
-CAMLprim value lwt_unix_bytes_read(value fd, value buf, value vofs, value vlen)
-{
-    intnat ofs, len, written;
-    DWORD numbytes, numwritten;
-    DWORD err = 0;
-
-    Begin_root(buf);
-    ofs = Long_val(vofs);
-    len = Long_val(vlen);
-    written = 0;
-    if (len > 0) {
-        numbytes = len;
-        if (Descr_kind_val(fd) == KIND_SOCKET) {
-            int ret;
-            SOCKET s = Socket_val(fd);
-            ret = recv(s, (char *)Caml_ba_array_val(buf)->data + ofs, numbytes,
-                       0);
-            if (ret == SOCKET_ERROR) err = WSAGetLastError();
-            numwritten = ret;
-        } else {
-            HANDLE h = Handle_val(fd);
-            if (!ReadFile(h, (char *)Caml_ba_array_val(buf)->data + ofs,
-                          numbytes, &numwritten, NULL))
-                err = GetLastError();
-        }
-        if (err) {
-            win32_maperr(err);
-            uerror("write", Nothing);
-        }
-        written = numwritten;
-    }
-    End_roots();
-    return Val_long(written);
-}
-
-/* +-----------------------------------------------------------------+
-   | Memory mapped files                                             |
-   +-----------------------------------------------------------------+ */
-
-CAMLprim value lwt_unix_get_page_size(value Unit)
-{
-    SYSTEM_INFO si;
-    GetSystemInfo(&si);
-    return Val_long(si.dwPageSize);
-}
-
 /* +-----------------------------------------------------------------+
    | JOB: read                                                       |
    +-----------------------------------------------------------------+ */
@@ -534,12 +387,8 @@ LWT_NOT_AVAILABLE3(unix_bytes_recv_msg)
 LWT_NOT_AVAILABLE3(unix_bytes_send_msg)
 LWT_NOT_AVAILABLE6(unix_bytes_sendto)
 LWT_NOT_AVAILABLE6(unix_bytes_sendto_byte)
-LWT_NOT_AVAILABLE6(unix_madvise)
-LWT_NOT_AVAILABLE4(unix_mincore)
 LWT_NOT_AVAILABLE2(unix_wait_mincore_job)
 LWT_NOT_AVAILABLE1(unix_guess_blocking_job)
-LWT_NOT_AVAILABLE1(unix_readable)
-LWT_NOT_AVAILABLE1(unix_writable)
 LWT_NOT_AVAILABLE1(unix_open_job)
 LWT_NOT_AVAILABLE1(unix_stat_job)
 LWT_NOT_AVAILABLE1(unix_lstat_job)
