+++ src/unix/unix_c/unix_bytes_read.c
@@ -1,0 +1,42 @@
+/* OCaml promise library
+ * http://www.ocsigen.org/lwt
+ * Copyright (C) 2009-2010 Jérémie Dimino
+ *               2009 Mauricio Fernandez
+ *               2010 Pierre Chambart
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, with linking exceptions;
+ * either version 2.1 of the License, or (at your option) any later
+ * version. See COPYING file for details.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA
+ * 02111-1307, USA.
+ */
+
+#include "lwt_config.h"
+
+#if !defined(LWT_ON_WINDOWS)
+
+#include <caml/mlvalues.h>
+#include <caml/unixsupport.h>
+#include <caml/bigarray.h>
+
+CAMLprim value lwt_unix_bytes_read(value val_fd, value val_buf, value val_ofs,
+                                   value val_len)
+{
+    long ret;
+    ret = read(Int_val(val_fd),
+               (char *)Caml_ba_array_val(val_buf)->data + Long_val(val_ofs),
+               Long_val(val_len));
+    if (ret == -1) uerror("read", Nothing);
+    return Val_long(ret);
+}
+#endif
